Adding the Drink and Patron classes
The Drink class holds the state of any mixed drink. It also contains some functions for getting the mixed flavor as well as mixed color of the drink.
Note: the `GetColor` function has a few blend modes that need to be tested.

The Patron class is a monobehavior in this case to allow for serializing of fields. The Patron class contains a flavor profile, List of Preferred and Hated Toppings.

Note the FlavorProfile is not showing up appropriately in the inspector at this time.